backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "copilot-swe-agent[bot]", email = "198982749+Copilot@users.noreply.github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "altcha>=1.0.0",
    "argon2-cffi>=23.1.0",
    "clamd>=1.0.2",
    "fastapi[standard]>=0.135.1",
    "passlib>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.37",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.6",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "S", "B", "A", "C4", "DTZ", "T20", "ICN", "PIE", "PT", "RSE", "RET", "SLF", "SIM", "TID", "TC", "ARG", "PTH", "ERA"]
ignore = ["S101", "S603", "S607", "B008", "S104", "S105"]

[tool.ruff.lint.per-file-ignores]
"alembic/*" = ["ERA001", "TC"]

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"
